updated the suggested changes by @arkid15r 1.deleted the badge queries - which is just querying all badges (not useful) 2.delete the test case for it 3.merge user_badge_field_test.py to user_test.py 4.updated user_badge.py to query for getting badges for each user
